Searchbar 컴포넌트 분리 + Typography 컴포넌트 추가
