+++ src/App.js
@@ -1,9 +1,7 @@
-import React from 'react';
-import Router from './router'
+import React from "react";
+import Router from "./router";
 function App() {
-  return (
-    <Router/>
-  );
+  return <Router />;
 }
 
 export default App;
